RAnEnExtra::plotAnalogSelection

RAnEnExtra::plotAnalogSelection visualize the selection of analog ensemble members. It shows how ensemble members are selected based on the similarity metrics. It also shows the selected forecasts for each parameter used during the analog generation, compared with the other unselected forecasts.

plotAnalogSelection(
  forecasts,
  fcst.times,
  sims,
  sims.index,
  test.times,
  search.times,
  i.station,
  i.test.day,
  i.flt,
  parameter.names,
  num.analogs,
  weights,
  cex.lab = 1.5,
  pch.selected = 16,
  pch.unselected = 16,
  pch.current = 1,
  col.selected = "red",
  col.unselected = "lightgrey",
  col.reference = "black",
  col.current = "black",
  lty.reference = "longdash",
  lwd.reference = 2,
  as.POSIXct = T,
  origin = "1970-01-01",
  tz = "UTC",
  single.figure = T,
  mar = c(1.5, 5, 1, 1) + 0.1,
  omi = c(0.3, 0, 0, 0),
  use.plotly = F,
  hovermode = "compare",
  spikemode = "across",
  spikethickness = 1,
  spikedash = "solid",
  spikecolor = "grey"
)

Arguments

forecasts
The forecasts data array

fcst.times
Forecast times

sims
The similarity member from analogs.

sims.index
The similarity index member from analogs.

i.station
The station index from analogs.

i.test.day
The test day index from analogs.

i.flt
The FLT index from analogs.

parameter.names
The parameter names that are associated with forecasts.

num.analogs
Number of analogs

weights
The weights used for AnEn

cex.lab
The font size of labels.

pch.selected
The point type for selected points.

pch.unselected
The point type for unselected points.

pch.current
The point type for current forecasts.

col.selected
The color for selected points.

col.unselected
The color for unselected points.

col.reference
The color for the reference line.

col.current
The color for current forecasts.

lty.reference
The line type for the reference line.

lwd.reference
The line width for the reference line.

as.POSIXct
Whether the time information should be converted to R date/time.

origin
The origin for times in forecasts. This will be passed to as.POSIXct to convert times in forecasts to date/time objects.

tz
The time zone in forecasts. This will be passed to as.POSIXct to convert times in forecasts to date/time objects.

single.figure
Whether to generate only a single figure.

mar
Margins for subplots. This is effective only when single.figure is TURE. It is passed to the function par.

omi
Outer margin for the plot. This is effective only when single.figure is TURE. It is passed to the function par.

use.plotly
Whether to use plotly. The plotly referrence can be found here.

hovermode
The hover mode for visualization. This is only effective when plotly is used.

spikemode
The spike mode for visualization. This is only effective when plotly is used.

spikethickness
The spike thickness for visualization. This is only effective when plotly is used.

spikedash
The spike dash style or the length in px for visualization. This is only effective when plotly is used.

spikecolor
The spike color for visualization. This is only effective when plotly is used.